refactor the tests a bit so it stoers the id rather than the model - DRYs up the code a bit

+++ src/sentry/workflow_engine/processors/workflow.py
@@ -170,7 +170,6 @@ def process_workflows(job: WorkflowJob) -> set[Workflow]:
 
 
 def delete_workflow(workflow: Workflow) -> bool:
-    # TODO - saponifi3d - add a sentry span here to measure the time it takes to delete a workflow
     with transaction.atomic(router.db_for_write(Workflow)):
         action_filters = DataConditionGroup.objects.filter(
             workflowdataconditiongroup__workflow=workflow
@@ -180,8 +179,13 @@ def delete_workflow(workflow: Workflow) -> bool:
             dataconditiongroupaction__condition_group__in=action_filters
         )
 
-        actions.delete()
-        action_filters.delete()
+        if actions:
+            actions.delete()
+
+        if action_filters:
+            action_filters.delete()
+
         workflow.when_condition_group.delete()
         workflow.delete()
+
     return True
